Adds edit modal and loads session data
